Changed publishedTime to int64.

+++ ddsenabler_participants/include/ddsenabler_participants/CBCallbacks.hpp
@@ -47,7 +47,7 @@ typedef void (*DdsNotification)(
         const char* typeName,
         const char* topicName,
         const char* json,
-        double publishTime);
+        int64_t publishTime);
 
 
 } /* namespace participants */
